+++ source/main/commands/import.js
@@ -14,7 +14,7 @@
 
 const ZettlrCommand = require('./zettlr-command')
 const { trans } = require('../../common/lang/i18n')
-const ZettlrImport = require('../zettlr-import')
+const makeImport = require('../modules/import')
 const path = require('path')
 
 class ImportFiles extends ZettlrCommand {
@@ -56,7 +56,7 @@ class ImportFiles extends ZettlrCommand {
     // Now import.
     global.ipc.notify(trans('system.import_status'))
     try {
-      let ret = ZettlrImport(fileList, this._app.getCurrentDir(), (file, error) => {
+      let ret = await makeImport(fileList, this._app.getCurrentDir(), (file, error) => {
         // This callback gets called whenever there is an error while running pandoc.
         global.ipc.notify(trans('system.import_error', path.basename(file)))
       }, (file) => {
@@ -71,6 +71,7 @@ class ImportFiles extends ZettlrCommand {
     } catch (e) {
       // There has been an error on importing (e.g. Pandoc was not found)
       // This catches this and displays it.
+      global.log.error(e.message, e)
       global.ipc.notify(e.message)
     }
   }