+++ source/main/modules/export/export.js
@@ -102,6 +102,11 @@ module.exports = async function (options) {
     case 'pdf':
       await preparePDF(options)
       break
+    case 'revealjs':
+      // Reveal.JS _always_ requires non-standalone
+      options.standalone = false
+      await prepareDefault(options)
+      break
     default:
       // In all cases, we want a standalone file
       // (docx and odt imply that, but RTF doesn't)