+++ source/main/zettlr-ql-standalone.js
@@ -72,6 +72,11 @@ class ZettlrQLStandalone {
       winConf.frame = false
     }
 
+    // Application icon for Linux. Cannot be not embedded in the executable.
+    if (process.platform === 'linux') {
+      winConf.icon = path.join(__dirname, 'assets/icons/128x128.png')
+    }
+
     // First create a new browserWindow
     let win = new BrowserWindow(winConf)