+++ source/renderer/assets/codemirror/zettlr-plugin-render-citations.js
@@ -16,7 +16,7 @@
   // blocks, second alternative are the simple @ID-things, both recognised by
   // Pandoc citeproc.
   // citationRE is taken from the Citr library (the extraction regex)
-  var citationRE = /(\[([^[\]]*@[^[\]]+)\])|(?<=\s|^)(@[\p{L}\d_][\p{L}\d_:.#$%&\-+?<>~/]*)/gu
+  var citationRE = /(\[(?:[^[\]]*@[^[\]]+)\])|(?<=\s|^)(@[\p{L}\d_][\p{L}\d_:.#$%&\-+?<>~/]*)/gu
   var citeMarkers = [] // CiteMarkers
   var currentDocID = null
   var Citr = require('@zettlr/citr')